+++ src/robot/parsing/__init__.py
@@ -49,6 +49,6 @@ Example
     print_suite(suite)
 """
 
-from .ast import ModelVisitor
-from .builders import get_ast, get_resource_ast
+from .builders import get_model, get_resource_model
 from .lexer import get_tokens, get_resource_tokens, Token
+from .model import ModelVisitor
